[Feature-58-products-calculated-price] some fixes
